Fix BLAS library compilation error in SuperLU
This commit addresses the compilation error encountered when building the
BLAS library within the SuperLU module. The problem was due to the absence
of the target directory for the libblas$(PLAT).a archive file. To resolve this,
we introduced a `directories` target in the Makefile to ensure the ../lib
directory exists before attempting to archive the BLAS object files.

The updated Makefile snippet now includes a rule to create the ../lib directory
before compiling and archiving the BLAS object files, preventing the "No such
file or directory" error during the `ar` command execution.

This fix resolves the build issue reported in https://github.com/JModelica/JModelica/issues/9.

fixes #9